Extract query logic into standalone files for library and series scenes
