+++ packages/browser/src/components/navigation/sidebar/sections/library/LibrarySideBarSection.tsx
@@ -8,7 +8,7 @@ import { useLocation } from 'react-router'
 
 import { useAppContext } from '@/context'
 import { usePaths } from '@/paths'
-import { usePrefetchLibrarySeries } from '@/scenes/library/tabs/series/LibrarySeriesScene'
+import { usePrefetchLibrarySeries } from '@/scenes/library/tabs/series/queries'
 
 import { EntityOptionProps } from '../../../types'
 import SideBarButtonLink from '../../SideBarButtonLink'
