+++ packages/browser/src/components/series/StackedSeriesCard.tsx
@@ -5,8 +5,8 @@ import { memo, useCallback, useEffect, useRef, useState } from 'react'
 import { Link } from '@/context'
 import { useFancyAnimations } from '@/hooks/useFancyAnimations'
 import { usePaths } from '@/paths'
-import { usePrefetchSeries } from '@/scenes/series'
-import { usePrefetchSeriesBooks } from '@/scenes/series/tabs/books/SeriesBooksScene'
+import { usePrefetchSeries } from '@/scenes/series/SeriesLayout'
+import { usePrefetchSeriesBooks } from '@/scenes/series/tabs/books/queries'
 
 import pluralizeStat from '../../utils/pluralize'
 import { SeriesStackedThumbnails } from '../thumbnail'
