+++ packages/browser/src/scenes/library/LibraryNavigation.tsx
@@ -9,8 +9,8 @@ import { useAppContext } from '@/context'
 import { usePreferences } from '@/hooks'
 
 import { useLibraryContext } from './context'
-import { usePrefetchLibraryBooks } from './tabs/books/LibraryBooksScene'
-import { usePrefetchLibrarySeries } from './tabs/series/LibrarySeriesScene'
+import { usePrefetchLibraryBooks } from './tabs/books/queries'
+import { usePrefetchLibrarySeries } from './tabs/series/queries'
 
 export default function LibraryNavigation() {
 	const location = useLocation()
