+++ packages/browser/src/scenes/series/tabs/books/queries.ts
@@ -1,0 +1,105 @@
+import { PREFETCH_STALE_TIME, useSDK } from '@stump/client'
+import { graphql, MediaFilterInput, MediaOrderBy } from '@stump/graphql'
+import { useQueryClient } from '@tanstack/react-query'
+import { useCallback } from 'react'
+
+import {
+	DEFAULT_MEDIA_ORDER_BY,
+	useSearchMediaFilter,
+	useURLKeywordSearch,
+	useURLPageParams,
+} from '@/components/filters/useFilterScene'
+
+export const query = graphql(`
+	query SeriesBooksScene(
+		$filter: MediaFilterInput!
+		$orderBy: [MediaOrderBy!]!
+		$pagination: Pagination!
+	) {
+		media(filter: $filter, orderBy: $orderBy, pagination: $pagination) {
+			nodes {
+				id
+				...BookCard
+				...BookMetadata
+			}
+			pageInfo {
+				__typename
+				... on OffsetPaginationInfo {
+					currentPage
+					totalPages
+					pageSize
+					pageOffset
+					zeroBased
+				}
+			}
+		}
+	}
+`)
+
+export type UsePrefetchSeriesBooksParams = {
+	page?: number
+	pageSize?: number
+	filter: MediaFilterInput[]
+	orderBy: MediaOrderBy[]
+}
+
+export function getQueryKey(
+	cacheKey: string,
+	libraryId: string,
+	page: number,
+	pageSize: number,
+	search: string | undefined,
+	filters: MediaFilterInput[] | undefined,
+	orderBy: MediaOrderBy[] | undefined,
+): (string | object | number | MediaFilterInput[] | MediaOrderBy[] | undefined)[] {
+	return [cacheKey, libraryId, page, pageSize, search, filters, orderBy]
+}
+
+export const usePrefetchSeriesBooks = () => {
+	const { sdk } = useSDK()
+	const { pageSize } = useURLPageParams()
+	const { search } = useURLKeywordSearch()
+	const searchFilter = useSearchMediaFilter(search)
+
+	const client = useQueryClient()
+
+	const prefetch = useCallback(
+		(
+			id: string,
+			params: UsePrefetchSeriesBooksParams = { filter: [], orderBy: DEFAULT_MEDIA_ORDER_BY },
+		) => {
+			const pageParams = { page: params.page || 1, pageSize: params.pageSize || pageSize }
+			return client.prefetchQuery({
+				queryKey: getQueryKey(
+					sdk.cacheKeys.seriesBooks,
+					id,
+					pageParams.page,
+					pageParams.pageSize,
+					search,
+					params.filter,
+					params.orderBy,
+				),
+				queryFn: async () => {
+					const response = await sdk.execute(query, {
+						filter: {
+							seriesId: { eq: id },
+							_and: params.filter,
+							_or: searchFilter,
+						},
+						orderBy: params.orderBy,
+						pagination: {
+							offset: {
+								...pageParams,
+							},
+						},
+					})
+					return response
+				},
+				staleTime: PREFETCH_STALE_TIME,
+			})
+		},
+		[client, search, searchFilter, pageSize, sdk],
+	)
+
+	return prefetch
+}
